ACTIN-4 Extract known genes from CKB and Hartwig
Introduces KnownGene to the serve datamodel and extractions, to be used as the canonical
list of cancer related genes with supporting evidence from our curated knowledgebases.

For CKB, each entry has a gene, so the list of known genes is simply the distinct set of these 
genes. The genes are annotated with roles from the variants.

For Hartwig, we extract genes from our driver gene panel and from an optional additional list
of manually curated genes.